Make ClientVersion its own module
