+++ ethcore/sync/src/chain/handler.rs
@@ -22,7 +22,8 @@ use ethcore::snapshot::{ManifestData, RestorationStatus};
 use ethcore::verification::queue::kind::blocks::Unverified;
 use ethereum_types::{H256, U256};
 use hash::keccak;
-use network::{ClientVersion, PeerId};
+use network::{PeerId};
+use network::client_version::ClientVersion;
 use rlp::Rlp;
 use snapshot::ChunkType;
 use std::time::Instant;
