+++ ethcore/sync/src/chain/requester.rs
@@ -18,7 +18,8 @@ use api::WARP_SYNC_PROTOCOL_ID;
 use block_sync::BlockRequest;
 use bytes::Bytes;
 use ethereum_types::H256;
-use network::{ClientCapabilities, ClientVersion, PeerId, PacketId};
+use network::{PeerId, PacketId};
+use network::client_version::{ClientCapabilities, ClientVersion};
 use rlp::RlpStream;
 use std::time::Instant;
 use sync_io::SyncIo;
